Port download script to python. Add documentation
The original script, written in Excel VBA, has been removed in favor for a
much more stable and quick script written in python 3. It also cleans the
course-list.txt file that it pulls from when a course no longer exists on
the uOttawa website. See README.md for more.

README.md updated with screenshots and more detailed operations.

courseCodes.csv updated with a refresh of the database made by the new
python script.

Added some fixes to ScheduleGeneratorServer: now quits when there is a
parsing error.